Split playlist.rs into engine clients and a server-test command
commands/playlist.rs mixed five unrelated subsystems across 3,400
lines. It now keeps only the open-playlist commands, load-progress
plumbing, and preview enrichment (916 lines); the rest moved verbatim:

- engine/xtream.rs: Xtream client (server normalizer, URL builders,
  account info, JSON-API playlist fetch)
- engine/stalker.rs: Stalker portal client
- engine/remote_cache.rs: ETag/Last-Modified download cache
- commands/server_test.rs: the Xtream server tester

Callers in saved.rs, recent.rs, and scan.rs keep addressing helpers via
crate::commands::playlist thanks to pub(crate) re-exports; unit tests
moved with their subjects.
