+++ src-tauri/src/engine/mod.rs
@@ -7,6 +7,9 @@ pub mod ffmpeg;
 pub mod parser;
 pub mod proxy;
 pub mod proxy_common;
+pub mod remote_cache;
 pub mod resume;
 pub mod search_pattern;
+pub mod stalker;
 pub mod stream_proxy;
+pub mod xtream;
